Improvement/improve notifications and internal data convention (#9)
* Removed all traces of API snakecase in views

Now the reducers make sure we discuss in camelCase consistently, and the actions know how to map this code. This reduces potential confusion and data convention inconsistent, which was plaging the workout form and details

* Cleaned up reducers/actions

Changes:
- No more 2 requests needed or duplicates to fetch all data
- fetchWorkoutAndExercises  now fires 2 events
- Reducers now get sent the minimal amount of data per type, interfacing with data is easier
- defaults being used to partially reset state on workout reducer to take care of inconsistent state
- workoutTemplate was using the wrong variable, this is now fixed. template loading should work much more proper now
- View model (selector) made for workoutform
- This seems to solve the exercise not loaded in form bug

* Fix the logout notification

The login form is changed to be a component, and a login container houses this form and the notification and the state. Now the notification is independent from the login form and this bug is finally fixed

* added newline at end on workout_selector